Allow using the same LDAP attribute for different config settings (#10437)
Some users want to use the same attribute (e.g. "cn") for different
config settings. (e.g. name and full name)

Also don't skip the unique ID attribute when collecting user attributes.
This is an issue when the same attribute is used for the unique
attribute and another config setting (note recommended) because the
attribute will be missing from user details.

Adjust the LDAP login test output to use a special key for the unique ID
attribute to avoid confusion because of the base64 encoded value.

Fixes #10069